+++ Game/Characters/Character.hpp
@@ -70,8 +70,8 @@ class Character {
 		std::vector<std::shared_ptr<Item>> items;
 		const bool isPlayerType;
 
-		int experiencePoints = 0;
-		int health = 200;
+		int_fast16_t experiencePoints = 0;
+		int_fast8_t health = 200;
 
 		std::shared_ptr<InputComponent> input;			//!< A smart pointer to an on the heap allocated InputComponent.
 		std::shared_ptr<PhysicsComponent> physics;		//!< A smart pointer to an on the heap allocated PhysicsComponent.
@@ -98,10 +98,10 @@ class Character {
 			return false;
 		}
 
-		int getExperience() const;
-		void setExperience(const int & experiencePointsToAdd);
-		int getHealth() const;
-		void setHealth(const int newHealth);
+		int_fast16_t getExperience() const;
+		void setExperience(const int_fast16_t & experiencePointsToAdd);
+		int_fast8_t getHealth() const;
+		void setHealth(const int_fast8_t newHealth);
 };
 
 #endif //__CHARACTER_HPP
