+++ Game/Items/Items.cpp
@@ -1,10 +1,10 @@
 #include "Items.hpp"
 
-Weapon::Weapon(const int damageFactor):
+Weapon::Weapon(const int_fast8_t damageFactor):
 	damageFactor(damageFactor)
 {}
 
-Consumable::Consumable(const int foodValue):
+Consumable::Consumable(const int_fast8_t foodValue):
 	foodValue(foodValue)
 {}
 
@@ -23,6 +23,6 @@ void Weapon::use(Character * character, std::vector<Character> & characters){
 void Consumable::use(Character * character, std::vector<Character> & characters){
 	character->setHealth(foodValue + character->getHealth());
 	if(character->getHealth() > 100){
-		character->setHealth(100);
+		character->setHealth(int_fast8_t(100));
 	}
 }
